parallel deletion works: port unordered_set to a phase concurrent hashtable.
Updates:
* vertex set is now a hashtable
* added a pair based hashtable that will be useful for the future.

+++ elektra/batch_dynamic_connectivity/graph.h
@@ -5,7 +5,7 @@
 #include <utility>
 
 /** Represents a vertex in a graph. */
-typedef int64_t Vertex;
+typedef int Vertex;
 
 /** Represents an edge in a directed graph. */
 typedef std::pair<Vertex, Vertex> DirectedEdge;
